Refactor planner store into typed modules
