+++ src/components/planner/PlannerPage.tsx
@@ -16,8 +16,8 @@ import WeekNotes from "./WeekNotes";
 import DayRow from "./DayRow";
 import ScrollTopFloatingButton from "./ScrollTopFloatingButton";
 import { useFocusDate, useWeek } from "./useFocusDate";
-import type { ISODate } from "./plannerStore";
-import { PlannerProvider } from "./plannerStore";
+import type { ISODate } from "./plannerTypes";
+import { PlannerProvider } from "./plannerContext";
 import WeekPicker from "./WeekPicker";
 import { PageHeader } from "@/components/ui";
 import PageShell from "@/components/ui/layout/PageShell";
