+++ src/components/planner/plannerTypes.ts
@@ -1,0 +1,35 @@
+export type ISODate = string;
+
+export type Project = {
+  id: string;
+  name: string;
+  done: boolean;
+  createdAt: number;
+  disabled?: boolean;
+  loading?: boolean;
+};
+
+export type DayTask = {
+  id: string;
+  title: string;
+  done: boolean;
+  projectId?: string;
+  createdAt: number;
+  images: string[];
+};
+
+export type DayRecord = {
+  projects: Project[];
+  tasks: DayTask[];
+  tasksById: Record<string, DayTask>;
+  tasksByProject: Record<string, string[]>;
+  doneCount: number;
+  totalCount: number;
+  focus?: string;
+  notes?: string;
+};
+
+export type Selection = {
+  projectId?: string;
+  taskId?: string;
+};
