+++ src/components/planner/useWeekData.ts
@@ -1,7 +1,8 @@
 "use client";
 
 import * as React from "react";
-import { ensureDay, type ISODate } from "./plannerStore";
+import { ensureDay } from "./plannerSerialization";
+import type { ISODate } from "./plannerTypes";
 import { usePlannerStore } from "./usePlannerStore";
 
 export function useWeekData(days: ISODate[]) {
